+++ tests/com/splunk/UtilTest.java
@@ -1,0 +1,73 @@
+/*
+ * Copyright 2012 Splunk, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"): you may
+ * not use this file except in compliance with the License. You may obtain
+ * a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations
+ * under the License.
+ */
+
+package com.splunk;
+
+import org.junit.Test;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public class UtilTest extends SplunkTestCase {
+    @Test
+    public void testJoin() {
+        List<String> emptyList = new ArrayList<String>();
+        SplunkTestCase.assertEquals("", Util.join("/", emptyList));
+
+        List<String> oneElementList = new ArrayList<String>();
+        oneElementList.add("abcd");
+        SplunkTestCase.assertEquals("abcd", Util.join("/", oneElementList));
+
+        List<String> fullList = new ArrayList<String>();
+        fullList.add("abcd");
+        fullList.add("defg");
+        SplunkTestCase.assertEquals(
+                "abcd/defg",
+                Util.join("/", fullList)
+        );
+    }
+
+    @Test
+    public void testJoinOnArray() {
+        String[] emptyArray = {};
+        SplunkTestCase.assertEquals("", Util.join("/", emptyArray));
+
+        String[] oneElementArray = {"abcd"};
+        SplunkTestCase.assertEquals("abcd", Util.join("/", oneElementArray));
+
+        String[] fullArray = {"abcd", "defg"};
+        SplunkTestCase.assertEquals(
+                "abcd/defg",
+                Util.join("/", fullArray)
+        );
+    }
+
+    @Test
+    public void testSubstringAfterSucceeds() {
+        SplunkTestCase.assertEquals(
+                "efg",
+                Util.substringAfter("abcdefg", "cd", "boris")
+        );
+    }
+
+    @Test
+    public void testSubstringAfterFails() {
+        SplunkTestCase.assertEquals(
+                "boris",
+                Util.substringAfter("abcdefg", "pq", "boris")
+        );
+    }
+}